move jsonschema out of function call parser
