+++ python/sglang/srt/function_call/function_call_parser.py
@@ -9,11 +9,11 @@ from sglang.srt.entrypoints.openai.protocol import (
 )
 from sglang.srt.function_call.base_format_detector import BaseFormatDetector
 from sglang.srt.function_call.core_types import ToolCallItem
+from sglang.srt.function_call.utils import _get_json_schema_constraint
 from sglang.srt.function_call.deepseekv3_detector import DeepSeekV3Detector
 from sglang.srt.function_call.deepseekv31_detector import DeepSeekV31Detector
 from sglang.srt.function_call.glm4_moe_detector import Glm4MoeDetector
 from sglang.srt.function_call.gpt_oss_detector import GptOssDetector
-from sglang.srt.function_call.json_array_detector import JsonArrayDetector
 from sglang.srt.function_call.kimik2_detector import KimiK2Detector
 from sglang.srt.function_call.llama32_detector import Llama32Detector
 from sglang.srt.function_call.mistral_detector import MistralDetector
@@ -46,7 +46,6 @@ class FunctionCallParser:
         "glm45": Glm4MoeDetector,
         "step3": Step3Detector,
         "gpt-oss": GptOssDetector,
-        "json": JsonArrayDetector,
     }
 
     def __init__(self, tools: List[Tool], tool_call_parser: str):
@@ -174,7 +173,7 @@ class FunctionCallParser:
             strict_tag = self.get_structure_tag()
             return ("structural_tag", strict_tag)
         elif tool_choice == "required" or isinstance(tool_choice, ToolChoice) or (isinstance(tool_choice, dict) and "function" in tool_choice):
-            json_schema = self.get_json_schema_constraint(tool_choice)
+            json_schema = _get_json_schema_constraint(self.tools, tool_choice)
             return ("json_schema", json_schema)
 
     def get_ebnf(
@@ -215,80 +214,3 @@ class FunctionCallParser:
 
         return self.detector.build_ebnf(filtered_tools)
 
-    def get_json_schema_constraint(
-        self, tool_choice: Union[ToolChoice, Literal["required"]]
-    ) -> Optional[dict]:
-        """
-        Get the JSON schema constraint for the specified tool choice.
-
-        Args:
-            tool_choice: The tool choice specification
-
-        Returns:
-            JSON schema dict, or None if no valid tools found
-        """
-        def get_tool_schema(tool):
-            return {
-                "properties": {
-                    "name": {
-                        "type": "string",
-                        "enum": [tool.function.name]
-                    },
-                    "parameters": tool.function.parameters
-                    if tool.function.parameters else {
-                        "type": "object",
-                        "properties": {}
-                    }
-                },
-                "required": ["name", "parameters"]
-            }
-        
-        if isinstance(tool_choice, ToolChoice) or (isinstance(tool_choice, dict) and "function" in tool_choice):
-            # For specific function choice, return the user's parameters schema directly
-            if isinstance(tool_choice, ToolChoice):
-                fn_name = tool_choice.function.name
-            else:
-                fn_name = tool_choice["function"]["name"]
-            for tool in self.tools:
-                if tool.function.name == fn_name:
-                    return {
-                        "type": "array",
-                        "minItems": 1,
-                        "maxItems": 1,
-                        "items": get_tool_schema(tool)
-                    }
-            return None
-        elif tool_choice == "required":
-
-            def get_tool_schema_defs(tools):
-                all_defs = {}
-                for tool in tools:
-                    if tool.function.parameters is None:
-                        continue
-                    # Make a copy to avoid modifying original
-                    params = tool.function.parameters.copy()
-                    defs = params.pop("$defs", {})
-                    for def_name, def_schema in defs.items():
-                        if def_name in all_defs and all_defs[def_name] != def_schema:
-                            raise ValueError(
-                                f"Tool definition '{def_name}' has "
-                                "multiple schemas, which is not "
-                                "supported.")
-                        else:
-                            all_defs[def_name] = def_schema
-                return all_defs
-
-            json_schema = {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "object",
-                    "anyOf": [get_tool_schema(tool) for tool in self.tools]
-                }
-            }
-            json_schema_defs = get_tool_schema_defs(self.tools)
-            if json_schema_defs:
-                json_schema["$defs"] = json_schema_defs
-            return json_schema
-
-        return None
